+++ lib/z_m3u_handler.dart
@@ -14,6 +14,8 @@ import 'package:z_m3u_handler/src/models/m3u_entry.dart';
 export 'package:z_m3u_handler/src/models/m3u_entry.dart';
 export 'package:z_m3u_handler/src/models/categorized_m3u_data.dart';
 export 'package:z_m3u_handler/src/models/classified_data.dart';
+export 'package:z_m3u_handler/src/models/credential_provider.dart';
+export 'package:z_m3u_handler/src/firebase/firebase_auth.dart';
 
 class ZM3UHandler {
   ZM3UHandler._pr();